Python 3.8 and latest alpine
This also changes so that we track the image digest to lock to a
specific image. Renovate will help us keep this updated. This also
has the advantage that we do not get surprises when building the code
some time later - only images that has successfully passed the tests
will be used.